Modificar cart controllers para filtro por usuario logueado Update y Remove

+++ src/controllers/cart.controllers.js
@@ -38,16 +38,18 @@ const create = catchError(async(req, res) => {
 
 const remove = catchError(async(req, res) => {
     const { id } = req.params;
-    const result = await Cart.destroy({ where: {id} });
+    const userId = req.user.id;
+    const result = await Cart.destroy({ where: {id, userId} });
     if(!result) return res.sendStatus(404);
     return res.sendStatus(204);
 });
 
 const update = catchError(async(req, res) => {
     const { id } = req.params;
+    const userId = req.user.id;
     const result = await Cart.update(
         req.body,
-        { where: {id}, returning: true }
+        { where: {id, userId}, returning: true }
     );
     if(result[0] === 0) return res.sendStatus(404);
     return res.json(result[1][0]);
